+++ src/util_graphics.h
@@ -47,14 +47,14 @@ extern ZGraphics z_graphics;
 
 extern void z_graphics_setup(void);
 
-extern void z_sprite_blit(ZSprite* Sprite, int8_t X, int8_t Y, uint8_t Frame);
-extern void z_sprite_blitCentered(ZSprite* Sprite, int8_t X, int8_t Y, uint8_t Frame);
+extern void z_sprite_blit(ZSprite* Sprite, int16_t X, int16_t Y, uint8_t Frame);
+extern void z_sprite_blitCentered(ZSprite* Sprite, int16_t X, int16_t Y, uint8_t Frame);
 extern int8_t z_sprite_getWidth(ZSprite* Sprite);
 extern int8_t z_sprite_getHeight(ZSprite* Sprite);
 
 extern void z_draw_fill(uint8_t Color);
-extern void z_draw_rectangle(int8_t X, int8_t Y, int8_t W, int8_t H, uint8_t Color);
-extern void z_draw_pixel(int8_t X, int8_t Y, uint8_t Color);
-extern void z_draw_circle(int8_t X, int8_t Y, uint8_t Radius, uint8_t Color);
+extern void z_draw_rectangle(int16_t X, int16_t Y, int8_t W, int8_t H, uint8_t Color);
+extern void z_draw_pixel(int16_t X, int16_t Y, uint8_t Color);
+extern void z_draw_circle(int16_t X, int16_t Y, uint8_t Radius, uint8_t Color);
 
 Z_EXTERN_C_END
